requirements.txt
# Core Dependencies
Django>=4.0,<5.0
djangorestframework>=3.12,<4.0
gunicorn>=20.1,<21.0
django-cors-headers>=3.7,<4.0
psycopg2-binary>=2.9,<3.0  # For PostgreSQL database
django-environ>=0.4,<1.0  # To handle environment variables
dj-database-url>=2.0,<3.0  # For database URL parsing
django-countries>=7.5,<8.0  # For country field support
django-enumfields>=2.1,<3.0  # For enum field support
django-filters>=23.0,<24.0  # For filtering support

# TensorFlow for AI/ML
tensorflow>=2.11,<3.0
scikit-learn>=1.0,<2.0
pandas>=1.3,<2.0
numpy>=1.21,<2.0
matplotlib>=3.5,<4.0

# Celery for background task processing
celery>=5.2,<6.0
redis>=4.5,<5.0

# WebSocket support
channels>=4.0,<5.0
channels-redis>=4.1,<5.0

# Caching
django-redis>=5.0,<6.0

# Security
djangorestframework-simplejwt>=5.0,<6.0  # For JWT Authentication

# Static files
whitenoise>=5.2,<6.0  # For serving static files in production

# Frontend Dependencies
django-webpack-loader>=1.9,<2.0  # If you’re using Webpack for frontend assets

# Testing and Linting
pytest>=7.0,<8.0
pytest-django>=4.4,<5.0
flake8>=5.0,<6.0
black>=22.0,<23.0
pylint>=2.13,<3.0

# Deployment
docker>=5.0,<6.0
